+++ okdata/sdk/data/download.py
@@ -7,6 +7,14 @@ from okdata.sdk.io_utils import write_file_content
 log = logging.getLogger()
 
 
+class DownloadURLAssertionError(Exception):
+    def __init__(self, url, expected_prefix):
+        super().__init__(
+            f"Aborting download because of an unexpected S3 download URL "
+            f"(should start with {expected_prefix}): {url}"
+        )
+
+
 class Download(SDK):
     def __init__(self, config=None, auth=None, env=None):
         self.__name__ = "download"
@@ -27,7 +35,13 @@ class Download(SDK):
         downloaded_files = []
         for file in self.get_files(dataset_id, version, edition, retries=retries):
             file_name = file["key"].split("/")[-1]
-            file_content_response = requests.get(file["url"], stream=True)
+            file_url = file["url"]
+            base_url = self.config.get("s3DownloadBaseUrl")
+
+            if not file_url.startswith(base_url):
+                raise DownloadURLAssertionError(file_url, base_url)
+
+            file_content_response = requests.get(file_url, stream=True)
             file_content_response.raise_for_status()
 
             write_file_content(file_name, output_path, file_content_response.raw.read())